% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_leroux.R
\name{load_leroux}
\alias{load_leroux}
\title{Load Leroux distribution}
\usage{
load_leroux()
}
\value{
\code{NULL}. The function defines \code{dcar_leroux} and \code{rcar_leroux} in the
global environment.
}
\description{
Define the Leroux CAR distribution functions used by NIMBLE.
}
\keyword{internal}
