% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_leroux.R
\name{load_leroux}
\alias{load_leroux}
\title{Load Leroux distribution}
\usage{
load_leroux()
}
\value{
No return value. The function assigns \code{dcar_leroux} and \code{rcar_leroux} to the
global environment.
}
\description{
Make the Leroux CAR distribution functions available in the global
environment so that NIMBLE can find them when processing BUGS model code.
}
\keyword{internal}
